% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_expr_type}
\alias{check_expr_type}
\title{check the expression type}
\usage{
check_expr_type(expr)
}
\description{
There are four types of expressions in R, including constant, symbol, call
and pairlist. This function was to check the type of expresison.
}
\keyword{internal}
