% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown_functions.R
\name{find_all_calls}
\alias{find_all_calls}
\title{find all calls in the expression}
\usage{
find_all_calls(expr)
}
\arguments{
\item{expr}{expression}
}
\description{
find all calls in the expression
}
\keyword{internal}
