% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown_symbols.R
\name{find_all_symbols}
\alias{find_all_symbols}
\title{Find all symbols in the given expression}
\usage{
find_all_symbols(expr)
}
\description{
Function used to find all symbols by walking the AST of R expression.
}
\keyword{internal}
