% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flat_map_chr}
\alias{flat_map_chr}
\title{flattern \code{map} results into a single character vector}
\usage{
flat_map_chr(.x, .f, ...)
}
\arguments{
\item{.x, .f, ...}{arguments inherits from \code{\link[purrr:map]{purrr::map()}}.}
}
\description{
flattern \code{map} results into a single character vector
}
\keyword{internal}
