% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.latexmath}
\alias{knit_print.latexmath}
\title{Custom knitr print methods for latexmath}
\usage{
\method{knit_print}{latexmath}(x, inline = FALSE, ...)
}
\arguments{
\item{x}{\code{latexmath} object to knit_print.}

\item{inline}{if the method is called in code chunks or inline R code}

\item{...}{Additional \code{\link[knitr:knit_print]{knitr::knit_print()}} arguments. Currently ignored,
except a optional argument \code{options}.}
}
\description{
The S3 method for print \code{latexmath} object in code chunks of knitr.
}
\references{
https://cran.r-project.org/web/packages/knitr/vignettes/knit_print.html#for-package-authors
}
\seealso{
\code{\link[knitr:knit_print]{knitr::knit_print()}}
}
