% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_env.R
\name{latex_env}
\alias{latex_env}
\title{Create an environment for latex math evaluation}
\usage{
latex_env(expr)
}
\arguments{
\item{expr}{an R expression.}
}
\description{
Create an environment for latex math evaluation
}
\keyword{internal}
