% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmath.R
\name{latexmath}
\alias{latexmath}
\title{Mark character as latex math}
\usage{
latexmath(char)
}
\arguments{
\item{char}{character}
}
\value{
The same value marked as \code{latexmath} class.
}
\description{
Making the giving character as \code{latexmath} object, which means the escape
method will not perform on it and leave it unchanged.
}
