% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_env.R
\name{make_env}
\alias{make_env}
\title{create environment for symbols or calls}
\usage{
make_env(.x, .f, parent = empty_env(), ...)
}
\arguments{
\item{.x}{a symbol or call.}

\item{.f}{function to transform symbols or calls to latex math.}

\item{...}{other arguments passed to \code{.f}.}
}
\value{
an environment.
}
\description{
create environment for symbols or calls
}
\keyword{internal}
