% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmath.R
\name{print.latexmath}
\alias{print.latexmath}
\title{Print method for latexmath}
\usage{
\method{print}{latexmath}(x, ...)
}
\arguments{
\item{x, }{\code{latexmath} to print.}

\item{..., }{other arguments passed to print.}
}
\description{
S3 method for printing \code{latexmath}, prints markup rmarkdown latex math
equations.
}
