% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlatexmath.R
\docType{package}
\name{rlatexmath}
\alias{rlatexmath}
\title{Generate latex math from R expressions.}
\description{
This package allows you to write latex mathematical equations from R
expressions directly.
}
