% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_env.R
\name{switch_expr}
\alias{switch_expr}
\title{switch evaluated type of \code{expr}}
\usage{
switch_expr(expr, ...)
}
\arguments{
\item{expr}{an expression.}

\item{...}{the list of alternatives.}
}
\description{
switch evaluated type of \code{expr}
}
\keyword{internal}
