% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_env.R
\name{transform_known_symbol}
\alias{transform_known_symbol}
\title{transform known symbols to latex math, add prefix \\}
\usage{
transform_known_symbol(symbol)
}
\description{
transform known symbols to latex math, add prefix \\
}
\keyword{internal}
