% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_symbols.R
\name{translate_known_symbol}
\alias{translate_known_symbol}
\title{transform known symbols to latex math, add prefix \\}
\usage{
translate_known_symbol(symbol)
}
\description{
transform known symbols to latex math, add prefix \\
}
\keyword{internal}
