% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_latexmath.R
\name{translate_latexmath}
\alias{translate_latexmath}
\alias{translate_latexmath_}
\title{Generate latex math in R}
\usage{
translate_latexmath(expr)

translate_latexmath_(expr)
}
\arguments{
\item{expr, }{R expression}
}
\description{
Transform R expression to latex math.
}
\examples{
translate_latexmath(pi)
translate_latexmath(alpha)

# keep the constant as is
translate_latexmath("pi")

# use translate_latexmath_ for quoted object
translate_latexmath_(quote(pi))

}
