% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown_functions.R
\name{translate_unknown_function}
\alias{translate_unknown_function}
\title{translate unknown functions}
\usage{
translate_unknown_function(op)
}
\arguments{
\item{op}{a operator.}
}
\description{
translate unknown functions
}
\keyword{internal}
