% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown_functions.R
\name{translate_unknown_functions}
\alias{translate_unknown_functions}
\title{multiple operators returned from \code{find_all_calls()}}
\usage{
translate_unknown_functions(calls)
}
\description{
multiple operators returned from \code{find_all_calls()}
}
\keyword{internal}
